raname and split commands support to inline commands and split file names correction and uploadasdoc thumbnail issue fix

+++ tobrot/helper_funcs/upload_to_tg.py
@@ -286,19 +286,12 @@ async def upload_single_file(message, local_file_name, caption_str, from_user, e
     #
     if UPLOAD_AS_DOC.upper() == 'TRUE' or dyna_user_config_upload_as_doc:
         thumb_image_path = None
-        if os.path.exists(thumbnail_location):
+        if thumbnail_location is not None and os.path.exists(thumbnail_location):
             thumb_image_path = await copy_file(
                 thumbnail_location,
                 os.path.dirname(os.path.abspath(local_file_name))
             )
-        if os.path.exists(thumb_image_path):
-            metadata = extractMetadata(createParser(thumb_image_path))
-            if metadata.has("width"):
-                width = metadata.get("width")
-            if metadata.has("height"):
-                height = metadata.get("height")
-            # ref: https://t.me/PyrogramChat/44663
-            # https://stackoverflow.com/a/21669827/4723940
+        if thumb_image_path is not None and os.path.exists(thumb_image_path):
             Image.open(thumb_image_path).convert(
                 "RGB"
             ).save(thumb_image_path)
